third_party/blink/web_tests/fast/dom/DOMImplementation/createHTMLDocument-optional-title-expected.txt

This tests optional title for createHTMLDocument

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.implementation.createHTMLDocument().querySelector("title") is null
PASS document.implementation.createHTMLDocument(undefined).querySelector("title") is null
PASS document.implementation.createHTMLDocument(null).querySelector("title").textContent is "null"
PASS document.implementation.createHTMLDocument("").querySelector("title").textContent is ""
PASS document.implementation.createHTMLDocument("x").querySelector("title").textContent is "x"
PASS successfullyParsed is true

TEST COMPLETE
